scripts/ml_training/requirements.txt
torch>=2.5.0
onnx>=1.15.0
onnx-simplifier>=0.4.35
scikit-learn>=1.4.0
pandas>=2.2.0
psycopg2-binary>=2.9.9
numpy>=1.26.0
tqdm>=4.66.0
python-dotenv>=1.0.0
